fix some numeric types cast issues
Former-commit-id: c199d9fdf2d806cc4d9d92725e6bcdcc51bd2334 [formerly 721aa3935dc8338494c60c00d7a4e48514f6f635]
Former-commit-id: 2f709f53af9d93697d4280906bf0a04fa3700bf0 [formerly c436a313a5ad32f678664acb681c81be94d3bed3]
Former-commit-id: d7e3dbdaccf1b6d1d0262f6d50e397c8c0d8e508

+++ plugins/bullet/src/minko/component/bullet/CylinderShape.hpp
@@ -102,7 +102,7 @@ namespace minko
 				float
 				volume() const
 				{
-					const float volume	= PI * (_halfExtentX + _margin) * (_halfExtentZ + _margin) // elliptical basis area approx
+					const float volume	= (float)PI * (_halfExtentX + _margin) * (_halfExtentZ + _margin) // elliptical basis area approx
 						* (_halfExtentY + _margin);
 
 					return volume * _volumeScaling;
